Added Pattern editor settings for various variables including the ability to turn on and off both the syntactic and the semantic highlights. Renamed TextHighlighter to the more appropriate  IdentifierHighlighter to make the distinction between this and the simpler preexisting syntactic highlighting code more clear. Also added whitespace, auto-indent, tab size and code folding settings.
This ends the changes to this PR which should be merged this weekend at the latest.

+++ plugins/ui/source/ui/text_editor/editor.cpp
@@ -413,7 +413,7 @@ namespace hex::ui {
                             if (index == std::string::npos)
                                 index = line.size() - 1;
                             if (index == 0) continue;
-                            u64 spacesToRemove = (index % m_tabSize) ? (index % m_tabSize) : m_tabSize;
+                            u64 spacesToRemove = (index % (u64) m_tabSize) ? (index % (u64) m_tabSize) : (u64) m_tabSize;
                             spacesToRemove = std::min(spacesToRemove, line.size());
                             line.erase(line.begin(), spacesToRemove);
                             line.m_colorized = false;
